Create README.md
This is a Power BI report using data from sales, orders, order details, customer details, product types, etc. to tell a story about what is happening in the company and most importantly what is profitable and what isn't. 

The report easily paints the picture through the dashboards what is hot and what to do next.